# Conflicts:
#	internal/adc/translator/httproute.go
#	internal/controller/utils.go

+++ internal/adc/translator/grpcroute.go
@@ -205,7 +205,7 @@ func (t *Translator) TranslateGRPCRoute(tctx *provider.TranslateContext, grpcRou
 				kind = string(*backend.Kind)
 			}
 			if backend.Port != nil {
-				port = int32(*backend.Port)
+				port = *backend.Port
 			}
 			namespace := string(*backend.Namespace)
 			name := string(backend.Name)
